requirements.txt
# Core
streamlit==1.38.0
python-dotenv==1.0.1

# LangChain ecosystem
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-chroma==0.1.3
langchain-huggingface==0.0.3
langchain-google-genai

# Google AI
google-generativeai

# Vector DB
chromadb==0.4.24

# PDF processing
pypdf==3.17.4

# Web scraping (NEW)
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

tqdm==4.66.4
